Rename prepared_metadata_extension IT to result_...
The extension does not affect prepared metadata (info about bind
markers), only result metadata.
