+++ scylla/tests/integration/statements/mod.rs
@@ -4,8 +4,8 @@ mod coordinator;
 mod execution_profiles;
 mod named_bind_markers;
 mod prepared;
-mod prepared_metadata_extension;
 mod request_timeout;
+mod result_metadata_extension;
 mod timestamps;
 mod transparent_reprepare;
 mod unprepared;
